initial empty commit to an orphan branch
I have a blog that is built with the 11ty static site generator. I want to merge --no-ff into main from the original repo of the theme when it's updated, so that it's a single merge commit of an update, and my history is completely separated from the history of the theme.

Otherwise it's a mix of my commits and the commits from the original theme's repo.

I'll write an elaborate blog post on this, when I have time.